COMPILE READY - Changed Implementation Strategy
Removed indexItems, indexCategories function because it creates extra documents
in the index. Created a general indexKeywords function that holds all of the
indexing. The function implements both the basic and advanced portions of the
search indexing.
Solved the issue of category indexing by concatenating the categories into one
category string that is passed to the indexer. The concatenation is implemented
with a HashMap.